setup yarn and jest
