+++ jest.config.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  collectCoverage: false,
+  collectCoverageFrom: ['./src/**/*.ts'],
+  coverageReporters: ['json', 'lcov'],
+  clearMocks: true
+}
