+++ packages/participants/src/messages/invalid_participant_evt.ts
@@ -4,29 +4,27 @@
 
 "use strict";
 
-import {DomainEventMsg} from "../../shared/domain_abstractions/messages";
+import {DomainEventMsg} from "shared";
 import {ParticipantsAggTopics} from "../domain/participants_agg";
 
 
-export class NetCapLimitExceededEvt extends DomainEventMsg{
+export class InvalidParticipantEvt extends DomainEventMsg{
 	aggregate_id: string;
 	aggregate_name: string = "Participants";
 	msg_key: string;
 	msg_topic: string = ParticipantsAggTopics.DomainEvents;
 
 	payload: {
-		participant_id:string;
-		transfer_id:string;
+		participant_id:string
 	};
 
-	constructor(participant_id:string, transfer_id:string) {
+	constructor(participant_id:string) {
 		super();
 
 		this.aggregate_id = this.msg_key = participant_id;
 
 		this.payload = {
-			participant_id,
-			transfer_id
+			participant_id
 		}
 	}
 }
