+++ packages/participants/test/example.test.ts
@@ -1,0 +1,7 @@
+describe('Example test', () => {
+
+  test('example test', async () => {
+    expect(1).toBe(1)
+  })
+
+})
