+++ packages/shared/src/domain_abstractions/base_aggregate.ts
