+++ packages/shared/src/domain_abstractions/base_entity_state.ts
