+++ packages/shared/src/domain_abstractions/ientity_state_repository.ts
