+++ packages/shared/src/index.ts
@@ -1,0 +1,15 @@
+// domain abstractions
+export * from './domain_abstractions/base_aggregate'
+export * from './domain_abstractions/base_entity_state'
+export * from './domain_abstractions/base_entity'
+export * from './domain_abstractions/entity_factory'
+export * from './domain_abstractions/ientity_state_repository'
+export * from './domain_abstractions/imessage_publisher'
+export * from './domain_abstractions/messages'
+
+// infrastructure
+export * from './infrastructure/kafka_generic_producer'
+export * from './infrastructure/kafka_message_publisher'
+
+// utilities
+export * from './utilities/logger'
