+++ packages/shared/src/infrastructure/kafka_generic_producer.ts
