+++ packages/shared/src/infrastructure/kafka_message_publisher.ts
